+++ src/app/[locale]/(protected)/_components/map/map-skeleton.tsx
@@ -1,23 +1,33 @@
+"use client";
+
 import { PiMapDuoSolid } from "@/components/icons";
-import { getScopedI18n } from "@/locales/server";
+import { cn } from "@/lib/utils";
+import { useScopedI18n } from "@/locales/client";
 
-interface MapSkeletonProps {}
+interface MapSkeletonProps {
+  isMapLoading: boolean;
+}
 
-export async function MapSkeleton({}: MapSkeletonProps) {
-  const t = await getScopedI18n("map.loading");
+export function MapSkeleton({ isMapLoading }: MapSkeletonProps) {
+  const t = useScopedI18n("map.loading");
 
   return (
-    <div className="absolute inset-0 flex animate-pulse items-center justify-center bg-muted dark:bg-border">
+    <figure
+      className={cn(
+        "relative flex flex-1 items-center justify-center bg-background/75 backdrop-blur-md transition-all delay-1000 duration-1000",
+        isMapLoading ? "visible z-20 opacity-100" : "invisible -z-10 opacity-0",
+      )}
+    >
       <div className="flex max-w-sm flex-col items-center gap-0.5">
-        <PiMapDuoSolid className="mb-4 h-10 w-10 text-muted-foreground" />
+        <PiMapDuoSolid className="mb-4 h-10 w-10 animate-pulse text-muted-foreground" />
 
-        <span className="text-balance text-center text-xl font-semibold text-accent-foreground">
+        <span className="text-balance text-center text-xl font-semibold text-muted-foreground">
           {t("title")}
         </span>
-        <span className="text-balance text-center text-sm text-muted-foreground">
+        <span className="text-balance text-center text-sm text-muted-foreground opacity-75">
           {t("subtitle")}
         </span>
       </div>
-    </div>
+    </figure>
   );
 }