+++ src/app/[locale]/(protected)/flights/[flightId]/page.tsx
@@ -3,7 +3,6 @@ import { NetworkIcon } from "@/components/network-icon";
 import { Skeleton } from "@/components/ui/skeleton";
 import { getFlightDetails } from "@/lib/flights";
 import { getScopedI18n } from "@/locales/server";
-import { Metadata, ResolvingMetadata } from "next";
 import dynamic from "next/dynamic";
 import { notFound } from "next/navigation";
 import { AircraftCard } from "./_components/aircraft-card";
@@ -12,6 +11,7 @@ import { BlurBackdrop } from "./_components/blur-backdrop";
 import { CrewCard } from "./_components/crew-card";
 import { FlightPlanDetails } from "./_components/flight-plan-details";
 import { FlightTelemetry } from "./_components/flight-telemetry";
+import { FlyToPlaneButton } from "./_components/fly-to-plane-button";
 const VerticalProfileChart = dynamic(
   () => import("./_components/vertical-profile-chart"),
   {
@@ -23,17 +23,6 @@ const VerticalProfileChart = dynamic(
 interface FlightsDetailPageProps {
   params: {
     flightId: string;
-  };
-}
-
-export async function generateMetadata(
-  { params: { flightId } }: FlightsDetailPageProps,
-  parent: ResolvingMetadata,
-): Promise<Metadata> {
-  const data = await getFlightDetails(flightId);
-
-  return {
-    title: data?.callsign,
   };
 }
 
@@ -62,8 +51,7 @@ export default async function FlightsDetailPage({
           <div className="flex flex-col">
             <h1 className="text-3xl font-bold">{data.callsign}</h1>
             <span className="text-sm text-muted-foreground">
-              {data.airline?.callsign || t("unknownCallsign")} &bull;{" "}
-              {data.airline?.name || t("unknownAirline")}
+              {data.airline?.callsign || t("unknownCallsign")}
             </span>
           </div>
         </div>
@@ -145,6 +133,7 @@ export default async function FlightsDetailPage({
       </section>
 
       <FlightPlanDetails data={data} />
+      <FlyToPlaneButton />
     </div>
   );
 }