+++ src/app/[locale]/(protected)/layout.tsx
@@ -1,6 +1,7 @@
-import { ReactNode, Suspense } from "react";
-import { MapContainer } from "./_components/map/map-container";
-import { MapSkeleton } from "./_components/map/map-skeleton";
+import { getProfile } from "@/lib/profile";
+import { ReactNode } from "react";
+import { Map } from "./_components/map";
+import { MapFooter } from "./_components/map/map-footer";
 import { MobileSidebar } from "./_components/mobile-sidebar";
 import { Sidebar } from "./_components/sidebar";
 import { Spotlight } from "./_components/spotlight";
@@ -9,20 +10,25 @@ interface AppRootLayoutProps {
   children: ReactNode;
 }
 
-export default function AppRootLayout({ children }: AppRootLayoutProps) {
+export default async function AppRootLayout({ children }: AppRootLayoutProps) {
+  const user = await getProfile();
+
   return (
     <div className="relative flex h-screen w-screen overflow-clip">
       <div className="flex w-full flex-1 flex-col justify-end lg:w-fit lg:flex-row lg:justify-normal">
-        <Sidebar />
+        <Sidebar user={user} />
         {children}
         <MobileSidebar />
       </div>
-
       <Spotlight />
-
-      <Suspense fallback={<MapSkeleton />}>
-        <MapContainer />
-      </Suspense>
+      <Map
+        userIntegrations={{
+          ivaoId: user.ivaoId,
+          vatsimId: user.vatsimId,
+        }}
+      >
+        <MapFooter />
+      </Map>
     </div>
   );
 }