+++ src/stores/base-map-store.ts
@@ -1,4 +1,5 @@
 import { create } from "zustand";
+import { persist } from "zustand/middleware";
 
 export type BaseMap = "light" | "dark" | "satellite" | "theme";
 
@@ -7,9 +8,16 @@ export interface BaseMapStore {
   setBaseMap: (baseMap: BaseMap) => void;
 }
 
-export const useBaseMapStore = create<BaseMapStore>((set) => {
-  return {
-    baseMap: "theme",
-    setBaseMap: (baseMap: BaseMap) => set({ baseMap }),
-  };
-});
+const DEFAULT_BASE_MAP: BaseMap = "theme";
+
+export const useBaseMapStore = create(
+  persist<BaseMapStore>(
+    (set) => ({
+      baseMap: DEFAULT_BASE_MAP,
+      setBaseMap: (baseMap: BaseMap) => set({ baseMap }),
+    }),
+    {
+      name: "base-map-store",
+    },
+  ),
+);